refactor: rename fs-tree workspace to arbor
