+++ .changeset/floppy-files-retire.md
@@ -1,0 +1,5 @@
+---
+"@jolly-pixel/arbor": minor
+---
+
+Rename fs-tree to arbor
