+++ ARCHITECTURE.md
@@ -12,7 +12,7 @@ packages/
   runtime/             @jolly-pixel/runtime              – Browser/Electron runtime (public)
   event-store/         @jolly-pixel/event-store          – Event Sourcing pattern applied to JollyPixel's events (public)
   network/             @jolly-pixel/network              – Shared multiplayer sync wire (public)
-  fs-tree/             @jolly-pixel/fs-tree              – Filesystem tree + live sync (public)
+  arbor/               @jolly-pixel/arbor                – Filesystem tree + live sync (public)
   fs-tree-backend/     @jolly-pixel/fs-tree-backend      – fs-tree socket server backend (private)
   resize-handle/       @jolly-pixel/resize-handle        – Resizable pane UI element (public)
   voxel-renderer/      @jolly-pixel/voxel.renderer       – Voxel rendering library (public)
