+++ README.md
@@ -27,7 +27,7 @@ Click on one of the links to access the documentation of the package:
 - [@jolly-pixel/network](./packages/network) - The shared wire for JollyPixel's multiplayer editors
 - [@jolly-pixel/voxel.renderer](./packages/voxel-renderer) - Voxel Engine and Renderer
 - [@jolly-pixel/pixel-draw.renderer](./packages/pixel-draw-renderer) - Pixel Art draw renderer 
-- [@jolly-pixel/fs-tree](./packages/fs-tree) - Robust, stylable tree view widget for HTML5 apps with drag'n'drop support
+- [@jolly-pixel/arbor](./packages/arbor) - Robust, stylable tree view widget for HTML5 apps with drag'n'drop support
 - [@jolly-pixel/resize-handle](./packages/resize-handle) - Robust resize handles / splitters for your HTML5 apps and websites
 - [@jolly-pixel/color](./packages/color) - The color utilities for JollyPixel's editors
