+++ packages/editors/voxel-model/package.json
@@ -14,8 +14,8 @@
     "dist"
   ],
   "dependencies": {
+    "@jolly-pixel/arbor": "1.0.1",
     "@jolly-pixel/engine": "4.0.0",
-    "@jolly-pixel/fs-tree": "1.0.1",
     "@jolly-pixel/pixel-draw.renderer": "3.0.0",
     "@jolly-pixel/resize-handle": "1.1.0",
     "@jolly-pixel/runtime": "4.0.0",
