+++ packages/fs-tree-backend/examples/index.html
@@ -19,7 +19,7 @@
   </main>
   <script src="https://cdn.socket.io/4.5.4/socket.io.min.js"></script>
   <script type="module">
-    import { TreeView } from "@jolly-pixel/fs-tree";
+    import { TreeView } from "@jolly-pixel/arbor";
 
     const socket = io("http://localhost:3000");
